# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Open Wallet API.",
        "title": "Open Wallet API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/transactions/entries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List a detailed view of entries joined with transactions for a given period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List entries",
                "operationId": "v1ListEntries",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter expression. \n- Allowed fields \u0026 ops:\n  - amount: eq, gt, gte, lt, lte\n  - category_id: eq, in\n  - created_at: eq, gt, gte, lt, lte\n  - id: eq, in\n  - period: eq, in, gte, lte\n  - reference_date: eq, gt, gte, lt, lte\n  - type: eq, in\n  - user_id: eq, in\n",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "reference_date:asc",
                        "description": "Sort field. \n- Allowed: reference_date, amount, id, created_at",
                        "name": "order_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of entries",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util.PaginatedResponse-internal_resources_transactions_handlers_ListEntriesResponseData"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns total income, expense and balance for each month in the specified period range.\nREQUIRED: period gte 'YYYYMM' and period lte 'YYYYMM' (max 12 months range)\nNote: Only periods with existing transactions/entries will be returned.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get financial summary grouped by month",
                "operationId": "v1GetSummary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter expression. \n- Allowed fields \u0026 ops:\n  - period: eq, in, gte, lte\n  - total_balance: eq, gt, gte, lt, lte\n  - total_expense: eq, gt, gte, lt, lte\n  - total_income: eq, gt, gte, lt, lte\n",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "period:asc",
                        "description": "Sort field. \n- Allowed: period, total_expense, total_income, total_balance",
                        "name": "order_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Summary data",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_transactions_handlers_SummaryResponseData"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/{transaction_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a transaction and all entries related by the ID of the transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Delete Transaction By ID",
                "operationId": "v1DeleteTransaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction ID",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transaction deleted"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Update a transaction",
                "operationId": "v1UpdateTransaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction ID",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Installment payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_resources_transactions_handlers.UpdateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Installment updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_transactions_handlers_UpdateTransactionResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_felipe1496_open-wallet_internal_util.PaginatedResponse-internal_resources_transactions_handlers_ListEntriesResponseData": {
            "type": "object",
            "required": [
                "data",
                "query"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_resources_transactions_handlers.ListEntriesResponseData"
                },
                "query": {
                    "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_querybuilder.Metadata"
                }
            }
        },
...
        "github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_transactions_handlers_SummaryResponseData": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_resources_transactions_handlers.SummaryResponseData"
                }
            }
        },
        "github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_transactions_handlers_UpdateTransactionResponseData": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_resources_transactions_handlers.UpdateTransactionResponseData"
                }
            }
        },
        "github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPErrorData"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "github_com_felipe1496_open-wallet_internal_util_httputil.HTTPErrorData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "github_com_felipe1496_open-wallet_internal_util_querybuilder.Metadata": {
            "type": "object",
            "required": [
                "next_page",
                "page",
                "per_page",
                "total_items",
                "total_pages"
            ],
            "properties": {
                "next_page": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "internal_resources_transactions_handlers.EntryResource": {
            "type": "object",
            "required": [
                "amount",
                "created_at",
                "id",
                "installment",
                "name",
                "period",
                "reference_date",
                "total_amount",
                "total_installments",
                "transaction_id",
                "type",
                "user_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category_color": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "installment": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "period": {
                    "type": "string"
                },
                "recurrence_id": {
                    "type": "string"
                },
                "reference_date": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "total_installments": {
                    "type": "integer"
                },
                "transaction_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_resources_transactions_handlers.ListEntriesResponseData": {
            "type": "object",
            "required": [
                "entries"
            ],
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_resources_transactions_handlers.EntryResource"
                    }
                }
            }
        },
        "internal_resources_transactions_handlers.MonthlySummaryResource": {
            "type": "object",
            "required": [
                "balance",
                "expense",
                "income",
                "period"
            ],
            "properties": {
                "balance": {
                    "type": "number"
                },
                "expense": {
                    "type": "number"
                },
                "income": {
                    "type": "number"
                },
                "period": {
                    "type": "string"
                }
            }
        },
        "internal_resources_transactions_handlers.SummaryResponseData": {
            "type": "object",
            "required": [
                "summary"
            ],
            "properties": {
                "summary": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_resources_transactions_handlers.MonthlySummaryResource"
                    }
                }
            }
        },
        "internal_resources_transactions_handlers.TransactionResource": {
            "type": "object",
            "required": [
                "created_at",
                "id",
                "name",
                "type",
                "user_id"
            ],
            "properties": {
                "category_color": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "recurrence_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_resources_transactions_handlers.UpdateEntryRequest": {
            "type": "object",
            "required": [
                "amount",
                "reference_date"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "maximum": 999999,
                    "minimum": -999999
                },
                "reference_date": {
                    "type": "string"
                }
            }
        },
        "internal_resources_transactions_handlers.UpdateTransactionRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "entries": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/internal_resources_transactions_handlers.UpdateEntryRequest"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "note": {
                    "type": "string",
                    "maxLength": 400,
                    "minLength": 0
                }
            }
        },
        "internal_resources_transactions_handlers.UpdateTransactionResponseData": {
            "type": "object",
            "required": [
                "transaction"
            ],
            "properties": {
                "transaction": {
                    "$ref": "#/definitions/internal_resources_transactions_handlers.TransactionResource"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}